# Optimize: ToJson.bytes for small inputs