`FromScala(true).transform(ToJson.bytes)` drops from 125ns => 105ns.

After this PR we:
- avoid an unnecessary array copy on `ByteArrayOutputStream.toByteArray`
- avoid creating a long chain of `Visitor.MapTo` instances.

I managed to drop the whole thing down to 85ns by:
- pushing the last `visitor.map` into `JsonGeneratorVisitor`, but it brought a lot of complexity
- creating my own BAOS that avoided the initial `new byte[](0)`.

I decided that last 20ns wasn't worth complicating the implementation or adding the amount of code required. I'll just take the easy 20% gain for now.

Most of the "Writing" benchmarks on https://plokhotnyuk.github.io/jsoniter-scala/ exercise this path, so we should see a small improvement there.

cc: @russellremple